# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pre-Order Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ── Orders ────────────────────────────────────────────────────────────────
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string, enum: [customer, admin, superadmin] }
        createdAt: { type: string, format: date-time }
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string, minLength: 8 }
        name: { type: string, minLength: 1 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }

    # ── Category ─────────────────────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [productId, productName, slug, unitPrice, currency, quantity, imageUrl]
      properties:
        productId: { type: string }
        productName: { type: string }
        slug: { type: string }
        unitPrice: { type: number }
        currency: { type: string }
        quantity: { type: integer }
        imageUrl: { type: ["string", "null"] }
        isPreOrder: { type: boolean }
    Cart:
      type: object
      required: [items, subtotal, currency, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
        variantSku: { type: string }
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 0 }

    # ── Orders ────────────────────────────────────────────────────────────────
...